tweak: Move manifests to separate files
