+++ radix-transactions/src/manifest/manifest_traits.rs
@@ -1,0 +1,25 @@
+use crate::internal_prelude::*;
+
+pub trait BuildableManifest: ReadableManifest + ManifestEncode + Default {
+    fn add_instruction(&mut self, instruction: Self::Instruction);
+    fn add_blob(&mut self, hash: Hash, content: Vec<u8>);
+    fn set_names(&mut self, names: KnownManifestObjectNames);
+}
+
+pub trait ReadableManifest {
+    type Instruction: ManifestInstruction;
+    fn get_instructions(&self) -> &[Self::Instruction];
+    fn get_blobs(&self) -> &IndexMap<Hash, Vec<u8>>;
+    fn get_preallocated_addresses(&self) -> &[PreAllocatedAddress] {
+        &NO_PREALLOCATED_ADDRESSES
+    }
+    fn get_child_subintents(&self) -> &[SubintentHash] {
+        &NO_CHILD_SUBINTENTS
+    }
+    fn get_known_object_names_ref(&self) -> ManifestObjectNamesRef;
+
+    fn validate(&self) -> Result<(), TransactionValidationError>;
+}
+
+static NO_PREALLOCATED_ADDRESSES: [PreAllocatedAddress; 0] = [];
+static NO_CHILD_SUBINTENTS: [SubintentHash; 0] = [];
