+++ radix-transactions/src/manifest/mod.rs
@@ -1,3 +1,4 @@
+mod any_manifest;
 pub mod ast;
 pub mod blob_provider;
 pub mod compiler;
@@ -8,13 +9,16 @@ pub mod dumper;
 pub mod e2e;
 pub mod generator;
 pub mod lexer;
-pub mod manifest_enums;
-pub mod manifest_naming;
+mod manifest_enums;
+mod manifest_naming;
+mod manifest_traits;
 pub mod parser;
 pub mod token;
 
+pub use any_manifest::*;
 pub use blob_provider::*;
 pub use compiler::{compile, CompileError};
 pub use decompiler::{decompile, decompile_any, DecompileError};
 pub use manifest_enums::*;
 pub use manifest_naming::*;
+pub use manifest_traits::*;
