+++ radix-transactions/src/model/v1/mod.rs
@@ -8,13 +8,14 @@ mod instruction_v1;
 mod instructions_v1;
 mod intent;
 mod intent_signatures;
-mod manifest;
+mod manifest_v1;
 mod message;
 mod notarized_transaction_v1;
 mod notary_signature;
 mod preview_transaction;
 mod round_update_transaction;
 mod signed_intent;
+mod system_manifest_v1;
 mod system_transaction;
 mod test_transaction;
 mod validated_notarized_transaction;
@@ -27,13 +28,14 @@ pub use instruction_v1::*;
 pub use instructions_v1::*;
 pub use intent::*;
 pub use intent_signatures::*;
-pub use manifest::*;
+pub use manifest_v1::*;
 pub use message::*;
 pub use notarized_transaction_v1::*;
 pub use notary_signature::*;
 pub use preview_transaction::*;
 pub use round_update_transaction::*;
 pub use signed_intent::*;
+pub use system_manifest_v1::*;
 pub use system_transaction::*;
 pub use test_transaction::*;
 pub use validated_notarized_transaction::*;
