+++ radix-transactions/src/model/v2/instructions_v2.rs
@@ -5,6 +5,18 @@ use crate::internal_prelude::*;
 #[sbor(transparent)]
 pub struct InstructionsV2(pub Rc<Vec<InstructionV2>>);
 
+impl From<Vec<InstructionV2>> for InstructionsV2 {
+    fn from(value: Vec<InstructionV2>) -> Self {
+        InstructionsV2(Rc::new(value))
+    }
+}
+
+impl From<InstructionsV2> for Vec<InstructionV2> {
+    fn from(value: InstructionsV2) -> Self {
+        value.0.as_ref().clone()
+    }
+}
+
 impl TransactionPartialPrepare for InstructionsV2 {
     type Prepared = PreparedInstructionsV2;
 }
