+++ radix-transactions/src/model/v2/mod.rs
@@ -6,6 +6,7 @@ mod instructions_v2;
 mod intent_core_v2;
 mod intent_header_v2;
 mod intent_signatures_v2;
+mod manifest_v2;
 mod message_v2;
 mod notarized_transaction_v2;
 mod partial_transaction_v2;
@@ -23,6 +24,7 @@ pub use instructions_v2::*;
 pub use intent_core_v2::*;
 pub use intent_header_v2::*;
 pub use intent_signatures_v2::*;
+pub use manifest_v2::*;
 pub use message_v2::*;
 pub use notarized_transaction_v2::*;
 pub use partial_transaction_v2::*;
